[MOD]-> Integer to Long

+++ src/main/java/model/repository/AirlineRepository.java
@@ -3,6 +3,6 @@ package model.repository;
 import model.entity.Airline;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AirlineRepository extends JpaRepository <Airline, Integer> {
+public interface AirlineRepository extends JpaRepository <Airline, Long> {
 
 }
