+++ src/main/java/model/repository/PassengerRepository.java
@@ -4,5 +4,5 @@ import model.entity.Passenger;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface PassengerRepository extends JpaRepository<Passenger, Integer> {
+public interface PassengerRepository extends JpaRepository<Passenger, Long> {
 }
